+++ packages/validation/src/validators/LionWebReferenceValidator.ts
@@ -22,9 +22,14 @@ export class LionWebReferenceValidator {
         this.validationResult = validationResult
     }
 
-    validateNodeIds(obj: LionWebJsonChunk, ctx: JsonContext): void {
+    /**
+     * Checks for duplicate node id's.
+     * As a side-effect creates a map from `id` to `node` for all nodes in `chunk` to make further validation faster.
+     * @param ctx
+     */
+    validateNodeIds(chunk: LionWebJsonChunk, ctx: JsonContext): void {
         // put all nodes in a map, validate that there are no two nodes with the same id.
-        obj.nodes.forEach((node, index) => {
+        chunk.nodes.forEach((node, index) => {
             // this.validationResult.check(this.nodesIdMap.get(node.id) === undefined, `Node number ${index} has duplicate id "${node.id}"`);
             if (!(this.nodesIdMap.get(node.id) === undefined)) {
                 this.validationResult.issue(new Reference_DuplicateNodeId_Issue(ctx.concat("nodes", index), node.id))
@@ -172,6 +177,13 @@ export class LionWebReferenceValidator {
         }
     }
 
+    /**
+     * Checks whether `child` appears in a `containment` of `parent`.
+     * Only applicable when both `parent` and `child` are in the same LionWebJsonChunk being validated.
+     * @param context
+     * @param parent
+     * @param child
+     */
     validateExistsAsChild(context: JsonContext, parent: LionWebJsonNode | undefined, child: LionWebJsonNode) {
         if (parent === undefined || parent === null) {
             return
@@ -187,6 +199,12 @@ export class LionWebReferenceValidator {
         this.validationResult.issue(new Reference_ChildMissingInParent_Issue(context, child, parent))
     }
 
+    /**
+     * Checks whether the `children` in the `containments` of `parent` have `parent` as their parent node.
+     * Only applicable when both `parent` and `child` are in the same LionWebJsonChunk being validated.
+     * @param node
+     * @param context
+     */
     validateChildrenHaveCorrectParent(node: LionWebJsonNode, context: JsonContext) {
         node.containments.forEach((child: LionWebJsonContainment) => {
             child.children.forEach((childId: LionWebId, index: number) => {