+++ packages/validation/src/validators/definitions/ResponseDefinitions.ts
@@ -1,0 +1,274 @@
+import { MessageGroup } from "../generic/schema/SyntaxDefinition.js";
+
+export const ResponseDefinitions: MessageGroup = {
+    name: "Response",
+    taggedUnionProperty: "messageKind",
+    sharedProperties: [
+        {
+            name: "messageKind",
+            type: "String",
+            isList: false,
+            isOptional: false,
+            mayBeNull: false,
+        },
+        {
+            name: "protocolMessages",
+            type: "ProtocolMessage",
+            isList: true,
+            isOptional: false,
+            mayBeNull: false,
+        },
+        {
+            name: "queryId",
+            type: "QueryId",
+            isList: false,
+            isOptional: false,
+            mayBeNull: false,
+        },
+    ],
+    messages: [
+        {
+            name: "SubscribeToChangingPartitionsResponse",
+            properties: [
+                {
+                    name: "messageKind",
+                    type: "String",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "protocolMessages",
+                    type: "ProtocolMessage",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "queryId",
+                    type: "QueryId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+            ],
+        },
+        {
+            name: "SubscribeToPartitionContentsResponse",
+            properties: [
+                {
+                    name: "contents",
+                    type: "LionWebDeltaJsonChunk",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "messageKind",
+                    type: "String",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "protocolMessages",
+                    type: "ProtocolMessage",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "queryId",
+                    type: "QueryId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+            ],
+        },
+        {
+            name: "UnsubscribeFromPartitionContentsResponse",
+            properties: [
+                {
+                    name: "messageKind",
+                    type: "String",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "protocolMessages",
+                    type: "ProtocolMessage",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "queryId",
+                    type: "QueryId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+            ],
+        },
+        {
+            name: "SignOnResponse",
+            properties: [
+                {
+                    name: "participationId",
+                    type: "ParticipationId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "messageKind",
+                    type: "String",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "protocolMessages",
+                    type: "ProtocolMessage",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "queryId",
+                    type: "QueryId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+            ],
+        },
+        {
+            name: "SignOffResponse",
+            properties: [
+                {
+                    name: "messageKind",
+                    type: "String",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "protocolMessages",
+                    type: "ProtocolMessage",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "queryId",
+                    type: "QueryId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+            ],
+        },
+        {
+            name: "ReconnectResponse",
+            properties: [
+                {
+                    name: "lastSentSequenceNumber",
+                    type: "SequenceNumber",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "messageKind",
+                    type: "String",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "protocolMessages",
+                    type: "ProtocolMessage",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "queryId",
+                    type: "QueryId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+            ],
+        },
+        {
+            name: "GetAvailableIdsResponse",
+            properties: [
+                {
+                    name: "ids",
+                    type: "LionWebId",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "messageKind",
+                    type: "String",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "protocolMessages",
+                    type: "ProtocolMessage",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "queryId",
+                    type: "QueryId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+            ],
+        },
+        {
+            name: "ListPartitionsResponse",
+            properties: [
+                {
+                    name: "partitions",
+                    type: "LionWebDeltaJsonChunk",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "messageKind",
+                    type: "String",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "protocolMessages",
+                    type: "ProtocolMessage",
+                    isList: true,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+                {
+                    name: "queryId",
+                    type: "QueryId",
+                    isList: false,
+                    isOptional: false,
+                    mayBeNull: false,
+                },
+            ],
+        },
+    ],
+};